Almost finished with WeatherMonitor
Figured out that JavaFX really doesn't like you to try to launch other
threads. Launching another window needs to happen on the same thread,
but it will stay until you close it. So WeatherMonitor has most of the
bones of its intended functionality, it can access the API, now it just
needs to be able to sort through what the state of the weather is, by my
making it check for the weather codes that the weather API sends, to see
if they are a certain set which we want to alert the user to. Right now,
I am only checking for weather that is currently happening, so checking
for extreme weather events that haven't happened yet would probably be
better, and I will do so probably.